refact(config): now config is only used to create objects not update
Before the config files could be used to update an object information, but that was leading to too many sanity checks and inconsistencies.

Now config files only create objects, if they ever get persisted to the database it's persistent version will prevent the creation of the non-persistent one.

Persistent objects should be changed in the database directly, if the server is off, or ingame with /code and other methods if the game is on.

Don't change the database directly with the server on, that could lead to inconsistencies.