refactor: classes receive only objects as arguments, code review
This update improves classes so they receive objects containing it's parameters as the only argument. This makes it easier to see the parameters of the object you are creating when you go back to it. Also you won't need to remember the order the parameters go when creating a new object.

The rest of the code had to be adapted for that, I also reviewed alot of stuff and fixed what I found to fix.

I'm making the code ready for new features I plan to implement.

This changes the style of some documentation.